moving isValidFileName into Utilities
